I caught the following snippet at the end of my /var/log/bsdinstall_log: === DEBUG: Running installation step: services local: Not in a function /usr/libexec/bsdinstall/services: cannot create  : Read-only file system /usr/libexec/bsdinstall/services: /tmp/bsdinstall/etc/rc.conf.services: \  Permission denied === The `local: Not in a function' is obvious, and was introduced by myself in SVN revision 256348.
The latter two are caused by the attempt to use "\" to continue the line
after using the ">>" redirect. This appears to attempt to write a file with
the name " " in the current directory and subsequently attempts to execute
the file that was originally intended for writing (which is not executable;
hence the `Permission denied'). That was introduced in SVN r228192 about
2 years ago, apparently unnoticed until I started going over the debug
outputs very carefully.

MFC after:	3 days