loader.efi: only fetch zfs pool guid for the actual boot device
With the zfs probe cleanup, the mistake did slip in the probe code;
instead of reading the pool GUID for the actual boot device (partition),
we read GUID for first found pool from the boot disk.

This will break the case when there are both zfs pool and ufs on the boot
disk, and the ufs is used for boot, not zfs.

Reviewed by:	smh
Differential Revision:	https://reviews.freebsd.org/D10359